Moved sattracks one level up as it will likely also be plotted on external figures.
